Added linkedin & email links to github profile readme
Also restructured a bit. The strange formatting where I have a <sub>text<sub> followed by a <p><sub>text</sub></p> is because this way the second text gets nested a bit closer to the first, compared to using a <p> tag on both, or a <br /> tag, both of which leave more spacing between the two pieces of text than my method.